Made a generalized SharedHeader component
